requirements.txt
pandas>=2.0.1
numpy>=1.23.4,<2.0.0
scikit-learn>=1.2.2,<2.0.0
scipy>=1.9.3,<2.0.0
joblib>=1.2.0,<2.0.0
Flask>=2.3.3,<3.0.0
python-dotenv>=1.0.1,<2.0.0
